rewrite in kotlin [close #15]
