+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ===
 
+Version 0.2.0 *(WIP)*
+---
+
+* Rewrite Plugin in Kotlin ([#15](https://github.com/yshrsmz/kgql/issues/15))
+
+
 Version 0.1.1 *(2019-02-05)*
 ---
 
@@ -10,28 +16,28 @@ Version 0.1.1 *(2019-02-05)*
 Version 0.1.0 *(2019-01-28)*
 ---
 
-* Generated Document Objects are now `internal` by default [#13](https://github.com/yshrsmz/kgql/issues/13)
+* Generated Document Objects are now `internal` by default ([#13](https://github.com/yshrsmz/kgql/issues/13))
 * Fix generated file's output directory not correct.
-* Kotlin is updated to 1.3.20 [#14](https://github.com/yshrsmz/kgql/issues/14)
-* Gradle is updated to 5.1.1 [#14](https://github.com/yshrsmz/kgql/issues/14) and __5.1.x or later__ is required.
+* Kotlin is updated to 1.3.20 ([#14](https://github.com/yshrsmz/kgql/issues/14))
+* Gradle is updated to 5.1.1 ([#14](https://github.com/yshrsmz/kgql/issues/14)) and __5.1.x or later__ is required.
 * Android Gradle Plugin is update to 3.3.0
 
 
 Version 0.0.7 *(2019-01-22)*
 ---
 
-* Use `kotlinx.serialization.SerialName` annotation [#11](https://github.com/yshrsmz/kgql/issues/11)
-* Android compilation task now depends on `generateKgqlInterface` task properly in Kotlin multiplatform project [#12](https://github.com/yshrsmz/kgql/issues/12)
+* Use `kotlinx.serialization.SerialName` annotation ([#11](https://github.com/yshrsmz/kgql/issues/11))
+* Android compilation task now depends on `generateKgqlInterface` task properly in Kotlin multiplatform project ([#12](https://github.com/yshrsmz/kgql/issues/12))
 * Gradle Plugin now depends on antlr4 to avoid dependency conflict with Android DataBinding
 
 
 Version 0.0.6 *(2019-01-21)*
 ---
 
-* Create dedicated object for each operation in a document([#3](https://github.com/yshrsmz/kgql/issues/3))
-* Change KgqlRequestBody & KgqlResponse to interface([#8](https://github.com/yshrsmz/kgql/issues/8))
+* Create dedicated object for each operation in a document ([#3](https://github.com/yshrsmz/kgql/issues/3))
+* Change KgqlRequestBody & KgqlResponse to interface ([#8](https://github.com/yshrsmz/kgql/issues/8))
 * Change suffix of generate classes to Document, from DocumentWrapper
-* Downgrade to Kotlin 1.3.11 [#8](https://github.com/yshrsmz/kgql/issues/8)
+* Downgrade to Kotlin 1.3.11 ([#8](https://github.com/yshrsmz/kgql/issues/8))
 
 
 Version 0.0.5 *(2019-01-15)*
