+++ kgql-compiler/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'org.jetbrains.kotlin.jvm'
-apply plugin: 'org.jetbrains.kotlin.kapt'
 
 sourceSets {
     main.java.srcDir "src/generated/kotlin"
@@ -14,9 +13,6 @@ dependencies {
     compile 'org.antlr:antlr4:4.7.1'
 
     implementation deps.kotlinPoet
-
-    implementation deps.moshi
-    kapt deps.moshiKotlinCodegen
 
     testImplementation deps.junit
     testImplementation deps.truth
