+++ kgql-gradle-plugin/build.gradle
@@ -1,7 +1,6 @@
-apply plugin: 'java-gradle-plugin'
-apply plugin: 'groovy'
-apply plugin: "com.gradle.plugin-publish"
 apply plugin: 'org.jetbrains.kotlin.jvm'
+apply plugin: 'java-gradle-plugin'
+apply plugin: "com.gradle.plugin-publish"
 
 sourceCompatibility = JavaVersion.VERSION_1_8
 targetCompatibility = JavaVersion.VERSION_1_8
@@ -24,7 +23,7 @@ gradlePlugin {
     plugins {
         kgql {
             id = "com.codingfeline.kgql"
-            implementationClass = "com.codingfeline.kgql.gradle.KgqlPlugin2"
+            implementationClass = "com.codingfeline.kgql.gradle.KgqlPlugin"
         }
     }
 }
@@ -43,15 +42,13 @@ tasks.getByName('pluginUnderTestMetadata')
 
 dependencies {
     implementation project(':kgql-compiler')
-//    implementation deps.kotlin.stdlib.jdk8
+    implementation deps.kotlin.stdlib.jdk8
 
     compileOnly gradleApi()
-    implementation localGroovy()
     implementation deps.plugins.kotlin
     implementation deps.plugins.serialization
     implementation deps.plugins.android
 
-    testImplementation deps.kotlin.stdlib.jdk8
     testImplementation deps.junit
     testImplementation deps.truth
 
@@ -65,16 +62,6 @@ compileKotlin {
 }
 compileTestKotlin {
     kotlinOptions.jvmTarget = "1.8"
-}
-
-compileGroovy {
-    sourceCompatibility = '1.8'
-    targetCompatibility = '1.8'
-}
-
-compileTestGroovy {
-    sourceCompatibility = '1.8'
-    targetCompatibility = '1.8'
 }
 
 test {
