--- kgql-gradle-plugin/src/main/groovy/com/codingfeline/kgql/gradle/KgqlExtension2.groovy
@@ -1,9 +1,0 @@
-package com.codingfeline.kgql.gradle
-
-import org.gradle.api.file.FileCollection
-
-class KgqlExtension2 {
-    String packageName = null
-    FileCollection sourceSet = null
-    Map<String, String> typeMapper = null
-}
