--- kgql-gradle-plugin/src/main/groovy/com/codingfeline/kgql/gradle/KgqlTask2.groovy
@@ -1,56 +1,0 @@
-package com.codingfeline.kgql.gradle
-
-import com.codingfeline.kgql.VersionKt
-import com.codingfeline.kgql.compiler.KgqlEnvironment
-import com.codingfeline.kgql.compiler.KgqlException
-import org.gradle.api.logging.LogLevel
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.OutputDirectory
-import org.gradle.api.tasks.SourceTask
-import org.gradle.api.tasks.TaskAction
-
-class KgqlTask2 extends SourceTask {
-    // Required to invalidate the task on version updates.
-    @Input
-    String pluginVersion() {
-        return VersionKt.getVERSION()
-    }
-
-    @OutputDirectory
-    File outputDirectory = null
-
-    @Input
-    Iterable<File> sourceFolders
-
-    @Input
-    String packageName
-
-    @Input
-    Map<String, String> typeMap
-
-    @TaskAction
-    void generateKgqlFiles() {
-        outputDirectory?.deleteDir()
-        outputDirectory.mkdirs()
-
-        def environment = new KgqlEnvironment(
-                getSource().toList(),
-                packageName,
-                outputDirectory,
-                typeMap
-        )
-
-        def generationStatus = environment.generateKgqlFiles { info -> logger.log(LogLevel.INFO, info) }
-
-        switch (generationStatus) {
-            case KgqlEnvironment.CompilationStatus.Failure:
-                logger.log(LogLevel.ERROR, "")
-                (generationStatus as KgqlEnvironment.CompilationStatus.Failure).errors.forEach {
-                    logger.log(LogLevel.ERROR, it)
-                }
-                throw new KgqlException("Generation failed; see the generator error output for details.")
-                break
-
-        }
-    }
-}
