--- plugin_bk/kotlin_bk/com/codingfeline/kgql/gradle/KgqlExtension.kt
@@ -1,9 +1,0 @@
-package com.codingfeline.kgql.gradle
-
-import org.gradle.api.file.FileCollection
-
-open class KgqlExtension {
-    var packageName: String? = null
-    var sourceSet: FileCollection? = null
-    var typeMapper: Map<String, String>? = null
-}
